Found a better way to do this.
Using the steam API was only going to capture games that people registered with steam. If I use the native discord presence I can capture every game. The plan is to store game data in Mongo and we could have a webpage that will display every single game being played. This will be something that runs on a repetated task to keep game data updated.

Discord does have a games tab but it's not really meant as a public style dashboard, it's personalized to your community and friends. This is going to be a public game browsers for things like LAN parties to help show what is the top games are.